+++ src/Eppie.App/Eppie.App.Shared/Services/ClipboardProvider.cs
@@ -1,19 +1,11 @@
-using Microsoft.Xaml.Interactivity;
 using System;
 using System.Threading.Tasks;
 using Tuvi.App.ViewModels.Services;
 using Windows.ApplicationModel.DataTransfer;
 
-#if WINDOWS_UWP
-using Windows.UI.Xaml.Controls;
-#else
-using Microsoft.UI.Xaml.Controls;
-#endif
-
-namespace Tuvi.App.Shared.Behaviors
+namespace Eppie.App.Shared.Services
 {
-    // ToDo: remove it [Eppie-io/Eppie-App#561]
-    public class ClipboardBehaviorBase : Behavior<Button>, IClipboardProvider
+    public class ClipboardProvider : IClipboardProvider
     {
         public async Task<string> GetClipboardContentAsync()
         {
