+++ src/Eppie.App/Eppie.App.Shared/Services/StorageItemsService.cs
@@ -1,16 +1,11 @@
-using System;
 using System.Collections.Generic;
-using System.Runtime.InteropServices.WindowsRuntime;
 using System.Threading.Tasks;
-using Tuvi.App.Shared.Behaviors;
-using Tuvi.App.ViewModels;
 using Tuvi.App.ViewModels.Services;
 using Windows.Storage;
-using Windows.Storage.Pickers;
 
-namespace Tuvi.App.Shared.Services
+namespace Eppie.App.Shared.Services
 {
-    public class StorageItemsService : FileBehavior
+    public class StorageItemsService : FileOperationProvider
     {
         private static IReadOnlyList<IStorageItem> Items { get; set; }
