+++ src/Eppie.App/Eppie.App.UI/Eppie.App.UI.Shared/Behaviors/CopyToClipboardBehavior.cs
@@ -8,6 +8,7 @@ using Microsoft.UI.Xaml;
 
 namespace Tuvi.App.Shared.Behaviors
 {
+    // ToDo: remove it [Eppie-io/Eppie-App#561]
     public class CopyToClipboardBehavior : ClipboardBehaviorBase
     {
         public static readonly DependencyProperty CopyCommandProperty =
