+++ src/Eppie.App/Eppie.App.UI/Eppie.App.UI.Shared/Behaviors/ItemClickCommandBehavior.cs
@@ -12,7 +12,7 @@ using Microsoft.UI.Xaml.Controls;
 
 namespace Tuvi.App.Shared.Behaviors
 {
-    // ToDo: it is not used and can be deleted
+    // ToDo: remove it [Eppie-io/Eppie-App#561]
     public class ItemClickCommandBehavior : Behavior<ListViewBase>
     {
         private bool _isItemClickEnabledCachedValue;
