+++ src/Eppie.App/Eppie.App.UI/Eppie.App.UI.Shared/Behaviors/ObjectToFileBehavior.cs
@@ -10,6 +10,7 @@ using Microsoft.UI.Xaml;
 
 namespace Tuvi.App.Shared.Behaviors
 {
+    // ToDo: remove it [Eppie-io/Eppie-App#561]
     public class AttachmentFileBehavior : ObjectToFileBehavior<Attachment> { }
 
     public class ObjectToFileBehavior<TObject> : FileBehavior
